# fix(tusb_msc): Cache sector metrics, enable storage buffering