@roma-jam 2025-02-10
@igi540 and please check the CI fail: `APP: Write data after disconnect`