@tore-espressif 2025-02-11
> @igi540 and please check the CI fail: `APP: Write data after disconnect`

This is really weird because it fails on v5.0 v5.1 and v5.2. 
It passes on v5.3, v5.4 and latest...