@roma-jam 2025-02-11
@igi540 Not it is not related, as in this error the espressif dcd layer is used: dcd_esp32sx.c:356

We are not using it since v0.15.