@tore-espressif 2025-02-11
@roma-jam sorry, I didn't provide any details. Very similar code with the same race condition and by-zero-division is also in the new DCD layer. However, I can't confirm it is the same problem now. Will look into it tomorrow