@tore-espressif 2025-04-09
@igi540 The problem with race condition is fixed https://github.com/espressif/tinyusb/pull/55

Please update this PR, so we can merge it after the fix is released